Add log space computations in all kernels. Add -inf initialization for algorithm's variables (alpha, beta, mu, f, count, ...) Add recover log possibility in print_grammar(..) method.

+++ include/algorithms/alg_inside_outside_main.h
@@ -8,9 +8,9 @@
 #include "kernels/update_parameters.cuh"
 #include "kernels/expect_count.cuh"
 
-long double* outside_algorithm(long double* mu, long double* beta, const uint32_t* sequence, 
+double* outside_algorithm(double* mu, double* beta, const uint32_t* sequence, 
     uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
                         #ifdef DEBUG_INSIDE_ALGORITHM
@@ -19,17 +19,17 @@ long double* outside_algorithm(long double* mu, long double* beta, const uint32_
 );
 
 
-long double* em_algorithm_calculate_expection_count(long double* count, long double* mu, long double* beta, 
+double* em_algorithm_calculate_expection_count(double* count, double* mu, double* beta, 
                         const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars
                         #ifdef DEBUG_INSIDE_ALGORITHM
                         ,pcfg* grammar
                         #endif
 );
 
-long double* inside_algorithm(const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
+double* inside_algorithm(const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path, pcfg* grammar);
